+++ src/lib/expressions/expressions.py
@@ -93,8 +93,10 @@ Networks
 
     Methods
 
-        s(<egress_port>,<ingress_port>) -> SParams
-            Gets an S-parameters.
+        s(<egress_port=None>,<ingress_port=None>,<rl_only=False>,<il_only=False>,<fwd_il_only=False>,<rev_il_only=False>) -> SParams
+            Returns S-parameters of a network.
+            <egress_port> and <ingress_port> can be set to a number, or kept at None (wildcard). Further filtering can be applied with
+            <rl_only>, <il_only>, <fwd_il_only>, <rev_il_only>.
 
         invert() -> Networks
             Inverts the ports (e.g. for de-embedding).